backend/requirements.txt
Django==2.2.16
pytest==6.2.4
psycopg2-binary==2.8.6 
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djoser==2.1.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
requests==2.26.
gunicorn==20.0.4
python-dotenv==0.19.2
drf-extra-fields==3.2.1
